# backend/artifacts/openapi/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/ops/slo-snapshot": {
      "get": {
        "operationId": "snapshot",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Aggregated technical and security metrics from PostgreSQL source of truth."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Read operational SLO snapshot for dashboards",
        "tags": [
          "Operational readiness"
        ]
      }
    },
...
    "/v1/owner/me": {
      "get": {
        "operationId": "profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Идентификатор, телефон и число питомцев."
          },
          "401": {
            "description": "Bearer token отсутствует или невалиден."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Профиль текущего владельца",
        "tags": [
          "Owner profile"
        ]
      }
    },
...
    "/v1/owner/bookings": {
      "get": {
        "operationId": "listBookings",
        "parameters": [
          {
            "name": "bucket",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "petId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Owner profile"
        ]
      }
    },
...
    "/v1/owner/bookings/{holdId}": {
      "get": {
        "operationId": "bookingDetail",
        "parameters": [
          {
            "name": "holdId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Owner profile"
        ]
      }
    },
...
    "/v1/clinic/{clinicId}/locations/{locationId}/patients/{patientId}/local-profile/reference": {
      "patch": {
        "operationId": "ClinicPatientAdministrativeReference_update",
        "parameters": [
          {
            "name": "clinicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "patientId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Correlation-ID",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "example": "\"0\""
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "administrativeReference"
                ],
                "properties": {
                  "administrativeReference": {
                    "type": "string",
                    "nullable": true,
                    "minLength": 1,
                    "maxLength": 40
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClinicPatientAdministrativeReferenceDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed headers or strict DTO."
          },
          "401": {
            "description": "Clinic employee JWT is required."
          },
          "403": {
            "description": "Mutation capability is required."
          },
          "404": {
            "description": "Rollout disabled or patient unavailable in the exact scope."
          },
          "409": {
            "description": "Stale version, collision, association change or key reuse."
          },
          "422": {
            "description": "Invalid administrative reference."
          },
          "428": {
            "description": "A strong If-Match is required."
          },
          "503": {
            "description": "Visibility policy is unavailable."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Set or clear an exact-scope clinic patient administrative reference",
        "tags": [
          "Clinic Patients"
        ],
        "x-required-capabilities": [
          "patient.admin.local-profile.update"
        ]
      }
    },
...
    "/v1/clinic/{clinicId}/locations/{locationId}/workspace-home": {
      "get": {
        "description": "Live database snapshot. Cache-Control: private, no-store. No ETag or 304 response is supported.",
        "operationId": "ClinicWorkspaceHome_read",
        "parameters": [
          {
            "name": "clinicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClinicWorkspaceHomeDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed clinicId or locationId."
          },
          "401": {
            "description": "Authentication required."
          },
          "403": {
            "description": "Normalized exact clinic/location membership denial."
          },
          "500": {
            "description": "Normalized technical failure."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Read the exact-location capability-filtered Clinic Workspace Home foundation",
        "tags": [
          "Clinic Workspace Home"
        ]
      }
    },
...
    "/v1/clinics": {
      "get": {
        "operationId": "listClinics",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "serviceCode",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "latitude",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "longitude",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "radiusKm",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "availableFrom",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "availableTo",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "openNow",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "telemedAvailable",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "emergencyCapability",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "selectedPetId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Активные клиники без внутренних MIS/provider полей."
          }
        },
        "summary": "Публичный список клиник",
        "tags": [
          "Public catalog"
        ]
      }
    },
    "/v1/clinics/{clinicId}": {
      "get": {
        "operationId": "readClinic",
        "parameters": [
          {
            "name": "clinicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Публичная карточка клиники",
        "tags": [
          "Public catalog"
        ]
      }
    },
...
    "/v1/owner/home": {
      "get": {
        "operationId": "OwnerHome_read",
        "parameters": [
          {
            "name": "selectedPetId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Minimal owned pet projections and one server-prioritized safe action."
          },
          "400": {
            "description": "selectedPetId is present but is not a UUID."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Owner-scoped home read model and next safe action",
        "tags": [
          "Owner home"
        ]
      }
    },
...
  },
  "info": {
    "title": "VetHelp Booking Core API",
    "description": "Контракт локального удержания слотов, подтверждения записи клиникой и освобождения hold.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "opaque Owner session or legacy staff JWT",
        "type": "http",
        "description": "Opaque vh_ Owner session or compatibility JWT for staff/system roles. Authorization and expiry are evaluated server-side for Owner sessions."
      }
    },
    "schemas": {
...
      "ClinicPatientAdministrativeReferenceDto": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "clinicId": {
            "type": "string",
            "format": "uuid"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "alias": {
            "type": "string",
            "nullable": true
          },
          "administrativeReference": {
            "type": "string",
            "nullable": true
          },
          "aggregateVersion": {
            "type": "number",
            "minimum": 1
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "patientId",
          "clinicId",
          "locationId",
          "alias",
          "administrativeReference",
          "aggregateVersion",
          "updatedAt"
        ]
      },
      "QueueWorkspaceFactsDto": {
        "type": "object",
        "properties": {
          "waitingCount": {
            "type": "number",
            "minimum": 0,
            "maximum": 999
          },
          "requiresActionCount": {
            "type": "number",
            "minimum": 0,
            "maximum": 999
          },
          "oldestWaitAgeBucket": {
            "type": "string",
            "enum": [
              "NONE",
              "LT_5_MIN",
              "5_TO_10_MIN",
              "10_TO_30_MIN",
              "GT_30_MIN"
            ]
          },
          "slaRisk": {
            "type": "string",
            "enum": [
              "NONE",
              "DUE_SOON",
              "OVERDUE",
              "UNKNOWN"
            ]
          }
        },
        "required": [
          "waitingCount",
          "requiresActionCount",
          "oldestWaitAgeBucket",
          "slaRisk"
        ]
      },
      "WorkspaceActionDto": {
        "type": "object",
        "properties": {
          "route": {
            "type": "string",
            "enum": [
              "queue",
              "appointments"
            ]
          },
          "labelKey": {
            "type": "string",
            "enum": [
              "WORKSPACE_OPEN_QUEUE",
              "WORKSPACE_OPEN_APPOINTMENTS"
            ]
          }
        },
        "required": [
          "route",
          "labelKey"
        ]
      },
      "QueueAvailableSectionDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "QUEUE"
            ]
          },
          "availability": {
            "type": "string",
            "enum": [
              "AVAILABLE"
            ]
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "facts": {
            "$ref": "#/components/schemas/QueueWorkspaceFactsDto"
          },
          "action": {
            "$ref": "#/components/schemas/WorkspaceActionDto"
          }
        },
        "required": [
          "kind",
          "availability",
          "generatedAt",
          "facts",
          "action"
        ]
      },
      "AppointmentsWorkspaceFactsDto": {
        "type": "object",
        "properties": {
          "todayCount": {
            "type": "number",
            "minimum": 0,
            "maximum": 999
          },
          "requiresActionCount": {
            "type": "number",
            "minimum": 0,
            "maximum": 999
          },
          "nextScheduledAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "todayCount",
          "requiresActionCount",
          "nextScheduledAt"
        ]
      },
      "AppointmentsAvailableSectionDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "APPOINTMENTS"
            ]
          },
          "availability": {
            "type": "string",
            "enum": [
              "AVAILABLE"
            ]
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "facts": {
            "$ref": "#/components/schemas/AppointmentsWorkspaceFactsDto"
          },
          "action": {
            "$ref": "#/components/schemas/WorkspaceActionDto"
          }
        },
        "required": [
          "kind",
          "availability",
          "generatedAt",
          "facts",
          "action"
        ]
      },
      "QueueUnavailableSectionDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "QUEUE"
            ]
          },
          "availability": {
            "type": "string",
            "enum": [
              "NOT_AUTHORIZED",
              "TEMPORARILY_UNAVAILABLE"
            ]
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Common database snapshot time; never a domain-source timestamp."
          }
        },
        "required": [
          "kind",
          "availability",
          "generatedAt"
        ]
      },
      "ScheduleUnavailableSectionDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "SCHEDULE"
            ]
          },
          "availability": {
            "type": "string",
            "enum": [
              "NOT_AUTHORIZED",
              "NOT_CONFIGURED"
            ]
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Common database snapshot time; never a domain-source timestamp."
          }
        },
        "required": [
          "kind",
          "availability",
          "generatedAt"
        ]
      },
      "AppointmentsUnavailableSectionDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "APPOINTMENTS"
            ]
          },
          "availability": {
            "type": "string",
            "enum": [
              "NOT_AUTHORIZED",
              "TEMPORARILY_UNAVAILABLE"
            ]
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Common database snapshot time; never a domain-source timestamp."
          }
        },
        "required": [
          "kind",
          "availability",
          "generatedAt"
        ]
      },
      "VeterinarianUnavailableSectionDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "VETERINARIAN"
            ]
          },
          "availability": {
            "type": "string",
            "enum": [
              "NOT_AUTHORIZED",
              "NOT_CONFIGURED"
            ]
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Common database snapshot time; never a domain-source timestamp."
          }
        },
        "required": [
          "kind",
          "availability",
          "generatedAt"
        ]
      },
      "QualityUnavailableSectionDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "QUALITY"
            ]
          },
          "availability": {
            "type": "string",
            "enum": [
              "NOT_AUTHORIZED",
              "NOT_CONFIGURED"
            ]
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Common database snapshot time; never a domain-source timestamp."
          }
        },
        "required": [
          "kind",
          "availability",
          "generatedAt"
        ]
      },
      "WorkspaceFreshnessDto": {
        "type": "object",
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "FRESH"
            ]
          },
          "maxAgeSeconds": {
            "type": "number",
            "enum": [
              30
            ]
          }
        },
        "required": [
          "state",
          "maxAgeSeconds"
        ]
      },
      "ClinicWorkspaceHomeDto": {
        "type": "object",
        "properties": {
          "clinicId": {
            "type": "string",
            "format": "uuid"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "serverNow": {
            "type": "string",
            "format": "date-time"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "freshness": {
            "$ref": "#/components/schemas/WorkspaceFreshnessDto"
          },
          "sections": {
            "type": "array",
            "minItems": 5,
            "maxItems": 5,
            "description": "Canonical order: QUEUE, SCHEDULE, APPOINTMENTS, VETERINARIAN, QUALITY.",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/QueueAvailableSectionDto"
                },
                {
                  "$ref": "#/components/schemas/QueueUnavailableSectionDto"
                },
                {
                  "$ref": "#/components/schemas/ScheduleUnavailableSectionDto"
                },
                {
                  "$ref": "#/components/schemas/AppointmentsAvailableSectionDto"
                },
                {
                  "$ref": "#/components/schemas/AppointmentsUnavailableSectionDto"
                },
                {
                  "$ref": "#/components/schemas/VeterinarianUnavailableSectionDto"
                },
                {
                  "$ref": "#/components/schemas/QualityUnavailableSectionDto"
                }
              ]
            }
          }
        },
        "required": [
          "clinicId",
          "locationId",
          "serverNow",
          "generatedAt",
          "freshness",
          "sections"
        ]
      },
...
    }
  }
}